storytrackpro/requirements.txt
django
mysqlclient
mysql-connector-python
djlint
python-dotenv
